feat(project): update to newman 5
Previously we had newman": "^3.5.2", but if you install exactly 3.5.2
e.g. with npm install newman@3.5.2 the ./api/run-api-tests.sh fails with:

Unrecognized arguments: --global-var

because that feature was added after 3.5. It works with npm install
because if fetches the newer version currently 3.10 which has it.

Since I was upgrading this, I though I might as well update to 5 as it just
works.