feat(project): make get favorited articles pai test unconditional
Similar rationale to the previous commit.

Both:

```
"name": "Articles Favorited by Username with auth",
"name": "Articles Favorited by Username",
```

are moved in the middle of:

The pre-existing get favorited tests were moved in between:

```
"name": "Favorite Article",
"name": "Unfavorite Article",
```

so that something will necessarily be favorited.

The tests used to check for a randomly named user `jane`, but this patch
modifies everything to be done under USERNAME. This way, we are certain
that the favorite is supposed to be present.